Updated lalinference_pipe
